Fix linking when PDAL is not installed in system library directory
The 'link_directories' command to specify the PDAL library directories
had no effect because it was specified after creating targets such as
align3d. This caused a link error when PDAL was not installed in a
system library directory.

This commit fixes the link error by calling link_directories before
creating targets.